LayoutTests/imported/w3c/web-platform-tests/svg/painting/animations/stroke-dasharray-composition-expected.txt


FAIL Compositing: property <stroke-dasharray> underlying [5 5] from add [0 5] to add [10 15] at (-0.6) should be [ 0  4] assert_equals: expected "0px , 4px " but got "0px , 0px "
FAIL Compositing: property <stroke-dasharray> underlying [5 5] from add [0 5] to add [10 15] at (-0.4) should be [ 1  6] assert_equals: expected "1px , 6px " but got "0px , 1px "
FAIL Compositing: property <stroke-dasharray> underlying [5 5] from add [0 5] to add [10 15] at (-0.2) should be [ 3  8] assert_equals: expected "3px , 8px " but got "0px , 3px "
FAIL Compositing: property <stroke-dasharray> underlying [5 5] from add [0 5] to add [10 15] at (0) should be [ 5 10] assert_equals: expected "5px , 10px " but got "0px , 5px "
FAIL Compositing: property <stroke-dasharray> underlying [5 5] from add [0 5] to add [10 15] at (0.2) should be [ 7 12] assert_equals: expected "7px , 12px " but got "2px , 7px "
FAIL Compositing: property <stroke-dasharray> underlying [5 5] from add [0 5] to add [10 15] at (0.4) should be [ 9 14] assert_equals: expected "9px , 14px " but got "4px , 9px "
FAIL Compositing: property <stroke-dasharray> underlying [5 5] from add [0 5] to add [10 15] at (0.6) should be [11 16] assert_equals: expected "11px , 16px " but got "6px , 11px "
FAIL Compositing: property <stroke-dasharray> underlying [5 5] from add [0 5] to add [10 15] at (0.8) should be [13 18] assert_equals: expected "13px , 18px " but got "8px , 13px "
FAIL Compositing: property <stroke-dasharray> underlying [5 5] from add [0 5] to add [10 15] at (1) should be [15 20] assert_equals: expected "15px , 20px " but got "10px , 15px "
FAIL Compositing: property <stroke-dasharray> underlying [5 5] from add [0 5] to add [10 15] at (1.2) should be [17 22] assert_equals: expected "17px , 22px " but got "12px , 17px "
FAIL Compositing: property <stroke-dasharray> underlying [10 10] from replace [5 10] to add [5 10] at (-0.6) should be [ 0  4] assert_equals: expected "0px , 4px " but got "5px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [10 10] from replace [5 10] to add [5 10] at (-0.4) should be [ 1  6] assert_equals: expected "1px , 6px " but got "5px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [10 10] from replace [5 10] to add [5 10] at (-0.2) should be [ 3  8] assert_equals: expected "3px , 8px " but got "5px , 10px "
PASS Compositing: property <stroke-dasharray> underlying [10 10] from replace [5 10] to add [5 10] at (0) should be [ 5 10]
FAIL Compositing: property <stroke-dasharray> underlying [10 10] from replace [5 10] to add [5 10] at (0.2) should be [ 7 12] assert_equals: expected "7px , 12px " but got "5px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [10 10] from replace [5 10] to add [5 10] at (0.4) should be [ 9 14] assert_equals: expected "9px , 14px " but got "5px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [10 10] from replace [5 10] to add [5 10] at (0.6) should be [11 16] assert_equals: expected "11px , 16px " but got "5px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [10 10] from replace [5 10] to add [5 10] at (0.8) should be [13 18] assert_equals: expected "13px , 18px " but got "5px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [10 10] from replace [5 10] to add [5 10] at (1) should be [15 20] assert_equals: expected "15px , 20px " but got "5px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [10 10] from replace [5 10] to add [5 10] at (1.2) should be [17 22] assert_equals: expected "17px , 22px " but got "5px , 10px "
PASS Compositing: property <stroke-dasharray> underlying [10 30] from add [none] to add [10 10] at (-1.5) should be [none]
PASS Compositing: property <stroke-dasharray> underlying [10 30] from add [none] to add [10 10] at (-0.5) should be [none]
PASS Compositing: property <stroke-dasharray> underlying [10 30] from add [none] to add [10 10] at (0) should be [none]
PASS Compositing: property <stroke-dasharray> underlying [10 30] from add [none] to add [10 10] at (0.4) should be [none]
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from add [none] to add [10 10] at (0.5) should be [20 40] assert_equals: expected "20px , 40px " but got "10px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from add [none] to add [10 10] at (0.6) should be [20 40] assert_equals: expected "20px , 40px " but got "10px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from add [none] to add [10 10] at (1) should be [20 40] assert_equals: expected "20px , 40px " but got "10px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from add [none] to add [10 10] at (1.5) should be [20 40] assert_equals: expected "20px , 40px " but got "10px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [0 5] from add [5] to replace [15 20 25] at (-0.2) should be [ 3  8  1  9  2  7] assert_equals: expected "3px , 8px , 1px , 9px , 2px , 7px " but got "3px , 2px , 1px "
FAIL Compositing: property <stroke-dasharray> underlying [0 5] from add [5] to replace [15 20 25] at (0) should be [ 5 10 5 10 5 10] assert_equals: expected "5px , 10px , 5px , 10px , 5px , 10px " but got "5px , 5px , 5px "
FAIL Compositing: property <stroke-dasharray> underlying [0 5] from add [5] to replace [15 20 25] at (0.2) should be [ 7 12  9 11  8 13] assert_equals: expected "7px , 12px , 9px , 11px , 8px , 13px " but got "7px , 8px , 9px "
FAIL Compositing: property <stroke-dasharray> underlying [0 5] from add [5] to replace [15 20 25] at (0.4) should be [ 9 14 13 12 11 16] assert_equals: expected "9px , 14px , 13px , 12px , 11px , 16px " but got "9px , 11px , 13px "
FAIL Compositing: property <stroke-dasharray> underlying [0 5] from add [5] to replace [15 20 25] at (0.6) should be [11 16 17 13 14 19] assert_equals: expected "11px , 16px , 17px , 13px , 14px , 19px " but got "11px , 14px , 17px "
FAIL Compositing: property <stroke-dasharray> underlying [0 5] from add [5] to replace [15 20 25] at (0.8) should be [13 18 21 14 17 22] assert_equals: expected "13px , 18px , 21px , 14px , 17px , 22px " but got "13px , 17px , 21px "
FAIL Compositing: property <stroke-dasharray> underlying [0 5] from add [5] to replace [15 20 25] at (1) should be [15 20 25 15 20 25] assert_equals: expected "15px , 20px , 25px , 15px , 20px , 25px " but got "15px , 20px , 25px "
FAIL Compositing: property <stroke-dasharray> underlying [0 5] from add [5] to replace [15 20 25] at (1.2) should be [17 22 29 16 23 28] assert_equals: expected "17px , 22px , 29px , 16px , 23px , 28px " but got "17px , 23px , 29px "
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from replace [5 10 15 20] to add [15 0 25 10 35 20] at (-0.2) should be [ 1  6 11 16  0  2 13 18  0  4  9 14] assert_equals: expected "1px , 6px , 11px , 16px , 0px , 2px , 13px , 18px , 0px , 4px , 9px , 14px " but got "3px , 12px , 13px , 22px , 0px , 8px , 15px , 24px , 1px , 10px , 11px , 20px , 3px , 12px , 13px , 22px , 0px , 8px , 15px , 24px , 1px , 10px , 11px , 20px "
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from replace [5 10 15 20] to add [15 0 25 10 35 20] at (0) should be [ 5 10 15 20  5 10 15 20  5 10 15 20] assert_equals: expected "5px , 10px , 15px , 20px , 5px , 10px , 15px , 20px , 5px , 10px , 15px , 20px " but got "5px , 10px , 15px , 20px , 5px , 10px , 15px , 20px , 5px , 10px , 15px , 20px , 5px , 10px , 15px , 20px , 5px , 10px , 15px , 20px , 5px , 10px , 15px , 20px "
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from replace [5 10 15 20] to add [15 0 25 10 35 20] at (0.2) should be [ 9 14 19 24 13 18 17 22 11 16 21 26] assert_equals: expected "9px , 14px , 19px , 24px , 13px , 18px , 17px , 22px , 11px , 16px , 21px , 26px " but got "7px , 8px , 17px , 18px , 11px , 12px , 15px , 16px , 9px , 10px , 19px , 20px , 7px , 8px , 17px , 18px , 11px , 12px , 15px , 16px , 9px , 10px , 19px , 20px "
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from replace [5 10 15 20] to add [15 0 25 10 35 20] at (0.4) should be [13 18 23 28 21 26 19 24 17 22 27 32] assert_equals: expected "13px , 18px , 23px , 28px , 21px , 26px , 19px , 24px , 17px , 22px , 27px , 32px " but got "9px , 6px , 19px , 16px , 17px , 14px , 15px , 12px , 13px , 10px , 23px , 20px , 9px , 6px , 19px , 16px , 17px , 14px , 15px , 12px , 13px , 10px , 23px , 20px "
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from replace [5 10 15 20] to add [15 0 25 10 35 20] at (0.6) should be [17 22 27 32 29 34 21 26 23 28 33 38] assert_equals: expected "17px , 22px , 27px , 32px , 29px , 34px , 21px , 26px , 23px , 28px , 33px , 38px " but got "11px , 4px , 21px , 14px , 23px , 16px , 15px , 8px , 17px , 10px , 27px , 20px , 11px , 4px , 21px , 14px , 23px , 16px , 15px , 8px , 17px , 10px , 27px , 20px "
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from replace [5 10 15 20] to add [15 0 25 10 35 20] at (0.8) should be [21 26 31 36 37 42 23 28 29 34 39 44] assert_equals: expected "21px , 26px , 31px , 36px , 37px , 42px , 23px , 28px , 29px , 34px , 39px , 44px " but got "13px , 2px , 23px , 12px , 29px , 18px , 15px , 4px , 21px , 10px , 31px , 20px , 13px , 2px , 23px , 12px , 29px , 18px , 15px , 4px , 21px , 10px , 31px , 20px "
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from replace [5 10 15 20] to add [15 0 25 10 35 20] at (1) should be [25 30 35 40 45 50 25 30 35 40 45 50] assert_equals: expected "25px , 30px , 35px , 40px , 45px , 50px , 25px , 30px , 35px , 40px , 45px , 50px " but got "15px , 0px , 25px , 10px , 35px , 20px , 15px , 0px , 25px , 10px , 35px , 20px , 15px , 0px , 25px , 10px , 35px , 20px , 15px , 0px , 25px , 10px , 35px , 20px "
FAIL Compositing: property <stroke-dasharray> underlying [10 30] from replace [5 10 15 20] to add [15 0 25 10 35 20] at (1.2) should be [29 34 39 44 53 58 27 32 41 46 51 56] assert_equals: expected "29px , 34px , 39px , 44px , 53px , 58px , 27px , 32px , 41px , 46px , 51px , 56px " but got "17px , 0px , 27px , 8px , 41px , 22px , 15px , 0px , 29px , 10px , 39px , 20px , 17px , 0px , 27px , 8px , 41px , 22px , 15px , 0px , 29px , 10px , 39px , 20px "
FAIL Compositing: property <stroke-dasharray> underlying [10 20] from replace [5 10 15] to add [10 5 20 15] at (-0.2) should be [ 2  7 12  0  8 13  0  5 14  1  6 11] assert_equals: expected "2px , 7px , 12px , 0px , 8px , 13px , 0px , 5px , 14px , 1px , 6px , 11px " but got "4px , 11px , 14px , 3px , 10px , 17px , 2px , 9px , 16px , 5px , 8px , 15px "
PASS Compositing: property <stroke-dasharray> underlying [10 20] from replace [5 10 15] to add [10 5 20 15] at (0) should be [ 5 10 15  5 10 15  5 10 15  5 10 15]
FAIL Compositing: property <stroke-dasharray> underlying [10 20] from replace [5 10 15] to add [10 5 20 15] at (0.2) should be [ 8 13 18 11 12 17 10 15 16  9 14 19] assert_equals: expected "8px , 13px , 18px , 11px , 12px , 17px , 10px , 15px , 16px , 9px , 14px , 19px " but got "6px , 9px , 16px , 7px , 10px , 13px , 8px , 11px , 14px , 5px , 12px , 15px "
FAIL Compositing: property <stroke-dasharray> underlying [10 20] from replace [5 10 15] to add [10 5 20 15] at (0.4) should be [11 16 21 17 14 19 15 20 17 13 18 23] assert_equals: expected "11px , 16px , 21px , 17px , 14px , 19px , 15px , 20px , 17px , 13px , 18px , 23px " but got "7px , 8px , 17px , 9px , 10px , 11px , 11px , 12px , 13px , 5px , 14px , 15px "
FAIL Compositing: property <stroke-dasharray> underlying [10 20] from replace [5 10 15] to add [10 5 20 15] at (0.6) should be [14 19 24 23 16 21 20 25 18 17 22 27] assert_equals: expected "14px , 19px , 24px , 23px , 16px , 21px , 20px , 25px , 18px , 17px , 22px , 27px " but got "8px , 7px , 18px , 11px , 10px , 9px , 14px , 13px , 12px , 5px , 16px , 15px "
FAIL Compositing: property <stroke-dasharray> underlying [10 20] from replace [5 10 15] to add [10 5 20 15] at (0.8) should be [17 22 27 29 18 23 25 30 19 21 26 31] assert_equals: expected "17px , 22px , 27px , 29px , 18px , 23px , 25px , 30px , 19px , 21px , 26px , 31px " but got "9px , 6px , 19px , 13px , 10px , 7px , 17px , 14px , 11px , 5px , 18px , 15px "
FAIL Compositing: property <stroke-dasharray> underlying [10 20] from replace [5 10 15] to add [10 5 20 15] at (1) should be [20 25 30 35 20 25 30 35 20 25 30 35] assert_equals: expected "20px , 25px , 30px , 35px , 20px , 25px , 30px , 35px , 20px , 25px , 30px , 35px " but got "10px , 5px , 20px , 15px , 10px , 5px , 20px , 15px , 10px , 5px , 20px , 15px "
FAIL Compositing: property <stroke-dasharray> underlying [10 20] from replace [5 10 15] to add [10 5 20 15] at (1.2) should be [23 28 33 41 22 27 35 40 21 29 34 39] assert_equals: expected "23px , 28px , 33px , 41px , 22px , 27px , 35px , 40px , 21px , 29px , 34px , 39px " but got "11px , 4px , 21px , 17px , 10px , 3px , 23px , 16px , 9px , 5px , 22px , 15px "
FAIL Compositing: property <stroke-dasharray> underlying [5] from add [0 5 10] to add [15 20 25 30 35] at (-0.2) should be [ 2  7 12  0  4 14  1  6 11  0  8 13  0  5 10] assert_equals: expected "2px , 7px , 12px , 0px , 4px , 14px , 1px , 6px , 11px , 0px , 8px , 13px , 0px , 5px , 10px " but got "0px , 2px , 7px , 0px , 0px , 9px , 0px , 1px , 6px , 0px , 3px , 8px , 0px , 0px , 5px "
FAIL Compositing: property <stroke-dasharray> underlying [5] from add [0 5 10] to add [15 20 25 30 35] at (0) should be [ 5 10 15  5 10 15  5 10 15  5 10 15  5 10 15] assert_equals: expected "5px , 10px , 15px , 5px , 10px , 15px , 5px , 10px , 15px , 5px , 10px , 15px , 5px , 10px , 15px " but got "0px , 5px , 10px , 0px , 5px , 10px , 0px , 5px , 10px , 0px , 5px , 10px , 0px , 5px , 10px "
FAIL Compositing: property <stroke-dasharray> underlying [5] from add [0 5 10] to add [15 20 25 30 35] at (0.2) should be [ 8 13 18 11 16 16  9 14 19 12 12 17 10 15 20] assert_equals: expected "8px , 13px , 18px , 11px , 16px , 16px , 9px , 14px , 19px , 12px , 12px , 17px , 10px , 15px , 20px " but got "3px , 8px , 13px , 6px , 11px , 11px , 4px , 9px , 14px , 7px , 7px , 12px , 5px , 10px , 15px "
FAIL Compositing: property <stroke-dasharray> underlying [5] from add [0 5 10] to add [15 20 25 30 35] at (0.4) should be [11 16 21 17 22 17 13 18 23 19 14 19 15 20 25] assert_equals: expected "11px , 16px , 21px , 17px , 22px , 17px , 13px , 18px , 23px , 19px , 14px , 19px , 15px , 20px , 25px " but got "6px , 11px , 16px , 12px , 17px , 12px , 8px , 13px , 18px , 14px , 9px , 14px , 10px , 15px , 20px "
FAIL Compositing: property <stroke-dasharray> underlying [5] from add [0 5 10] to add [15 20 25 30 35] at (0.6) should be [14 19 24 23 28 18 17 22 27 26 16 21 20 25 30] assert_equals: expected "14px , 19px , 24px , 23px , 28px , 18px , 17px , 22px , 27px , 26px , 16px , 21px , 20px , 25px , 30px " but got "9px , 14px , 19px , 18px , 23px , 13px , 12px , 17px , 22px , 21px , 11px , 16px , 15px , 20px , 25px "
FAIL Compositing: property <stroke-dasharray> underlying [5] from add [0 5 10] to add [15 20 25 30 35] at (0.8) should be [17 22 27 29 34 19 21 26 31 33 18 23 25 30 35] assert_equals: expected "17px , 22px , 27px , 29px , 34px , 19px , 21px , 26px , 31px , 33px , 18px , 23px , 25px , 30px , 35px " but got "12px , 17px , 22px , 24px , 29px , 14px , 16px , 21px , 26px , 28px , 13px , 18px , 20px , 25px , 30px "
FAIL Compositing: property <stroke-dasharray> underlying [5] from add [0 5 10] to add [15 20 25 30 35] at (1) should be [20 25 30 35 40 20 25 30 35 40 20 25 30 35 40] assert_equals: expected "20px , 25px , 30px , 35px , 40px , 20px , 25px , 30px , 35px , 40px , 20px , 25px , 30px , 35px , 40px " but got "15px , 20px , 25px , 30px , 35px , 15px , 20px , 25px , 30px , 35px , 15px , 20px , 25px , 30px , 35px "
FAIL Compositing: property <stroke-dasharray> underlying [5] from add [0 5 10] to add [15 20 25 30 35] at (1.2) should be [23 28 33 41 46 21 29 34 39 47 22 27 35 40 45] assert_equals: expected "23px , 28px , 33px , 41px , 46px , 21px , 29px , 34px , 39px , 47px , 22px , 27px , 35px , 40px , 45px " but got "18px , 23px , 28px , 36px , 41px , 16px , 24px , 29px , 34px , 42px , 17px , 22px , 30px , 35px , 40px "
FAIL Compositing: property <stroke-dasharray> underlying [1em] from add [9em 14px] to add [304px 1.5em] at (-0.2) should be [8em 28px] assert_equals: expected "128px , 28px " but got "112px , 12px "
FAIL Compositing: property <stroke-dasharray> underlying [1em] from add [9em 14px] to add [304px 1.5em] at (0) should be [10em 30px] assert_equals: expected "160px , 30px " but got "144px , 14px "
FAIL Compositing: property <stroke-dasharray> underlying [1em] from add [9em 14px] to add [304px 1.5em] at (0.2) should be [12em 32px] assert_equals: expected "192px , 32px " but got "176px , 16px "
FAIL Compositing: property <stroke-dasharray> underlying [1em] from add [9em 14px] to add [304px 1.5em] at (0.4) should be [14em 34px] assert_equals: expected "224px , 34px " but got "208px , 18px "
FAIL Compositing: property <stroke-dasharray> underlying [1em] from add [9em 14px] to add [304px 1.5em] at (0.6) should be [16em 36px] assert_equals: expected "256px , 36px " but got "240px , 20px "
FAIL Compositing: property <stroke-dasharray> underlying [1em] from add [9em 14px] to add [304px 1.5em] at (0.8) should be [18em 38px] assert_equals: expected "288px , 38px " but got "272px , 22px "
FAIL Compositing: property <stroke-dasharray> underlying [1em] from add [9em 14px] to add [304px 1.5em] at (1) should be [20em 40px] assert_equals: expected "320px , 40px " but got "304px , 24px "
FAIL Compositing: property <stroke-dasharray> underlying [1em] from add [9em 14px] to add [304px 1.5em] at (1.2) should be [22em 42px] assert_equals: expected "352px , 42px " but got "336px , 26px "
